tracker/scraper.py
import requests
from bs4 import BeautifulSoup

def scrape_daraz_product(url):
    # Implement scraping logic to get product data
    response = requests.get(url)
    soup = BeautifulSoup(response.content, 'html.parser')
    # Parse the necessary details
    product_data = {
        'name': 'Product Name',
        'daraz_id': 'UniqueID',
        'price': 100.00,
        'image_urls': ['http://image.url'],
        # ... other fields
    }
    return product_data

def scrape_daraz_reviews(url):
    # Implement scraping logic to get reviews
    reviews = []
    # Parse reviews and append to the list
    reviews.append({
        'reviewer_name': 'John Doe',
        'rating': 5,
        'content': 'Great product!',
        'date_posted': '2023-01-01',
    })
    return reviews
